Split main form methods into separate class file, sorted and split code into regions
