+++ Babel/Babel.csproj
@@ -121,6 +121,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="BabelMethods.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="frmWindowPicker.cs">
       <SubType>Form</SubType>
     </Compile>
